requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
streamlit-quill
sqlalchemy